Atualizações

+++ src/company/services/company.service.ts
@@ -2,30 +2,33 @@ import { InjectRepository } from "@nestjs/typeorm";
 import { Repository } from "typeorm";
 import { CompanyDTO } from "../dto/company.dto";
 import { Company } from "../entities/company.entity";
+import { User } from "src/user/entities/user.entity";
+import { Injectable } from "@nestjs/common";
 
+@Injectable()
 export class CompanyService {
     constructor(@InjectRepository(Company)
     private companyRepository: Repository<Company>) {}
 
-    async create(companyDTO: CompanyDTO): Promise<Company> {
-        return await this.companyRepository.save(companyDTO);
+    async create(user: User, companyDTO: CompanyDTO): Promise<Company> {
+        return await this.companyRepository.save({user_id: user.id, ...companyDTO});
     }
 
-    async update(id: number, companyDTO: CompanyDTO) {
-        await this.companyRepository.update({ id }, companyDTO);
+    async update(user: User, companyDTO: CompanyDTO) {
+        await this.companyRepository.update({ user_id: user.id }, companyDTO);
 
-        return await this.companyRepository.findOne({ where: { id }});
+        return this.findByUser(user);
     }
 
-    async remove(id: number): Promise<void> {
-        await this.companyRepository.delete(id);
+    async remove(user: User): Promise<void> {
+        await this.companyRepository.delete({ user_id: user.id });
     }
 
-    async findOne(id: number): Promise<Company> {
-        return await this.companyRepository.findOne({ where: { id } });
+    async findOne(user: User, id: number): Promise<Company> {
+        return await this.companyRepository.findOne({ where: { id, user_id: user.id } });
     }
 
-    async findAll(): Promise<Company[]> {
-        return await this.companyRepository.find();
+    async findByUser(user): Promise<Company> {
+        return await this.companyRepository.findOne({ where: { user_id: user.id } });
     }
 }
