+++ src/user_courses/services/user_courses.service.ts
@@ -3,6 +3,7 @@ import { InjectRepository } from "@nestjs/typeorm";
 import { UserCourses } from "../entities/user_courses.entity";
 import { Repository } from "typeorm";
 import { UserCoursesDTO } from "../dto/user_courses.dto";
+import { User } from "src/user/entities/user.entity";
 
 @Injectable()
 export class UserCoursesService {
@@ -11,25 +12,25 @@ export class UserCoursesService {
         private userCoursesRepository: Repository<UserCourses>
     ) {}
 
-    async create(userCoursesDTO: UserCoursesDTO): Promise<UserCourses> {
-        return await this.userCoursesRepository.save(userCoursesDTO);
+    async create(user: User, userCoursesDTO: UserCoursesDTO): Promise<UserCourses> {
+        return await this.userCoursesRepository.save({ user_id: user.id, ...userCoursesDTO});
     }
 
-    async update(id: number, userCoursesDTO: UserCoursesDTO): Promise<UserCourses> {
-        await this.userCoursesRepository.update({id}, userCoursesDTO);
+    async update(id: number, user: User, userCoursesDTO: UserCoursesDTO): Promise<UserCourses> {
+        await this.userCoursesRepository.update({id, user_id: user.id}, userCoursesDTO);
 
         return this.userCoursesRepository.findOne({ where: { id }});
     }
 
-    async remove(id: number): Promise<void> {
-        await this.userCoursesRepository.delete(id)
+    async remove(user: User, id: number): Promise<void> {
+        await this.userCoursesRepository.delete({ id, user_id: user.id })
     }
 
-    async findeOne(id: number): Promise<UserCourses> {
-        return this.userCoursesRepository.findOne({ where: { id }});
+    async findeOne(user: User, id: number): Promise<UserCourses> {
+        return this.userCoursesRepository.findOne({ where: { id, user_id: user.id }});
     }
 
-    async findAll(): Promise<UserCourses[]> {
-        return this.userCoursesRepository.find();
+    async findAll(user: User): Promise<UserCourses[]> {
+        return this.userCoursesRepository.find({ where: { user_id: user.id } })
     }
 }
